safari-534.53-branch/LayoutTests/http/tests/security/beforeload-iframe-server-redirect-expected.txt

CONSOLE MESSAGE: line 15: http://127.0.0.1:8000/security/resources/redir.php?url=http://localhost:8000/security/resources/post-done.html
This test produces output in the console. Only the initial URL should be logged. 
